api/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-seed==0.2.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Faker==4.1.3
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
pep8==1.7.1
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
